backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
pinecone
openai
fitz
PyMuPDF